specialty

+++ app/Http/Controllers/Dashboard/SpecialtyController.php
@@ -47,35 +47,52 @@ class SpecialtyController extends Controller
         return redirect()->route('specialties.index');
     }
 
-
     public function edit(Specialty $specialty)
     {
         return view('dashboard.specialties.edit', compact('specialty'));
     }
 
-    public function update(Request $request, Specialty $specialty)
+    public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'required'
-        ]);
-        $input = $request->except('_token', '_method');
-        $specialty->spec_icon = $input['spec_icon'];
-        $specialty->about = $input['about'];
-        $specialty->update($input);
-        return Redirect::route('specialties.edit', $specialty->id)->with('success', 'Done');
+        $specialty = Specialty::where('id', $id)->first();
+        if ($specialty) {
+            $request->validate([
+                "title" => "sometimes|required|unique:specialties,title," . $specialty->id,
+                "brief" => "sometimes|required|max:500",
+            ]);
+
+            $specialty->title = $request->title ?? $specialty->title;
+            if ($request->file('image')) {
+                $request->validate(["image" => "sometimes|image|mimes:jpeg,png,jpg|max:5128"]);
+                $specialty->image = $this->UploudAvatar($request->file('image'), 'specialities') ?? $specialty->image;
+            }
+            $specialty->brief = $request->brief ?? $specialty->brief;
+            $specialty->save();
+
+            notify()->success('You are awesome, Specialty has been Updated successfull!');
+            return redirect()->route('specialties.index');
+        } else {
+            notify()->success('Opps, Update Faild!');
+            return redirect()->back();
+        }
     }
 
     public function search(Request $request)
     {
         $search = $request->get('search');
         $allSpecialties = Specialty::where('name', 'like', '%' . $search . '%')->paginate(10);
-        return view('dashboard.specialties.all', compact('allSpecialties'))->with('SearchDone', 'Search Done');
+        return view('dashboard.specialties.manage', compact('allSpecialties'));
     }
 
-    public function destroy($id)
+    public function destroy(string $id)
     {
-        $specialty = Specialty::findOrFail($id);
-        $specialty->delete();
-        return Redirect::route('specialties.index')->with('success', 'Done');
+        $specialty = Specialty::find($id);
+        if ($specialty) {
+            $specialty->delete();
+            notify()->success('You are awesome, The Specialty has been deleted successfull!');
+        } else {
+            notify()->error('Opps!, The Specialty has been deleted before');
+        }
+        return redirect()->back();
     }
 }
